move tests for docdetails merge to `test_paper.py`
