+++ tests/test_paperqa.py
@@ -8,6 +8,7 @@ from collections.abc import AsyncIterable, Sequence
 from copy import deepcopy
 from io import BytesIO
 from pathlib import Path
+from datetime import datetime
 from typing import cast
 
 import httpx
@@ -1237,6 +1238,63 @@ def test_answer_rename(recwarn) -> None:
 def test_dois_resolve_to_correct_journals(doi_journals):
     details = DocDetails(doi=doi_journals["doi"])  # type: ignore[call-arg]
     assert details.journal == doi_journals["journal"]
+
+@pytest.mark.asyncio
+async def test_docdetails_merge_with_non_list_fields():
+    # Create two DocDetails instances
+    doc1 = DocDetails(
+        citation="Citation 1",
+        publication_date=datetime(2023, 1, 1),
+        docname="Document 1",
+        dockey="key1",
+        other={"bibtex_source": "source1", "client_source": ["client1"]},
+    )
+
+    doc2 = DocDetails(
+        citation="Citation 2",
+        publication_date=datetime(2024, 1, 1),
+        docname="Document 2",
+        dockey="key2",
+        other={"bibtex_source": ["source2"], "client_source": "client2"},
+    )
+
+    # Merge the two DocDetails instances
+    merged_doc = doc1 + doc2
+
+    # use set operations to check for containment
+    assert {"source1", "source2"}.issubset(merged_doc.other["bibtex_source"])
+    assert {"client1", "client2"}.issubset(merged_doc.other["client_source"])
+    # Check that the merged_doc is an instance of DocDetails
+    assert isinstance(merged_doc, DocDetails)
+
+
+@pytest.mark.asyncio
+def test_docdetails_merge_with_list_fields():
+    doc1 = DocDetails(
+        citation="Citation 1",
+        publication_date=datetime(2023, 1, 1),
+        docname="Document 1",
+        dockey="key1",
+        other={"bibtex_source": ["source1"], "client_source": ["client1"]},
+    )
+
+    doc2 = DocDetails(
+        citation="Citation 2",
+        publication_date=datetime(2024, 1, 1),
+        docname="Document 2",
+        dockey="key2",
+        other={"bibtex_source": ["source2"], "client_source": ["client2"]},
+    )
+
+    # Merge the two DocDetails instances
+    merged_doc = doc1 + doc2
+
+    # Check that 'other' fields are merged correctly
+    assert {"source1", "source2"}.issubset(merged_doc.other["bibtex_source"])
+    assert {"client1", "client2"}.issubset(merged_doc.other["client_source"])
+
+    # Check that the merged_doc is an instance of DocDetails
+    assert isinstance(merged_doc, DocDetails)
 
 
 @pytest.mark.vcr
